MH-13341, Deleting Capture Agents Should Not Modify Users
Right now, deleting a capture agent will delete the roles associated
with a capture agent from all existing users, regardless of the roles
having been assigned to that user before the agent was created or not.

This may lead to weird situations where user loose access when capture
agents are temporarily deleted but again added later on.

This patch removes this behavior, leaving the roles in place. This
should be no problem since the roles represent the unique identifiers of
the capture agents and adding a new agent with the same identifier
later on should mean that it is the exact same agent.