MH-13347, Don't always look for orphaned properties
For every delete query executed by the asset manager and regardless of
what is being deleted, the asset manager will look for any orphaned
asset manager properties (defined as properties for media packages for
which no snapshot exists) by joining the complete snapshot table with
the complete property table which is a quite expensive operation.

This patch will remove this procedure and instead add a simple,
performant cleanup procedure for a single media package before the first
snapshot is being created.