MH-12631 Drop the ORGANIZER field from the ical feed
This migrates ea4f738e66af51a2386e505064ac166c6f8d9d90 from 4.x to 6.x as it somehow got dropped,
probably as a result of the module rename process.

Original description:

Opencast includes ORGANIZER only if there is a Presenter (DC Creator) in the metadata.
So the lack of an ORGANIZER: field is not a problem to any existing CAs because Opencast allows
scheduling events without a presenter name.

Second, Opencast's use of the ORGANIZER field in the ical feed is not standards-compliant.

See Specifically https://www.ietf.org/rfc/rfc2445.txt, 4.8.4.3

From a standards point of view, it makes no sense to identify the presenter as the ical ORGANIZER
because the presenter is not sending a group scheduled event or publishing his/her busy time.

If we think of the CA as the "user" it also doesn't make much sense because we're not using VFREEBUSY.

The fake ORGANIZER format used by Opencast can cause problems with parsing the feed for some CAs,
especially when the presenter name includes quotes e.g.

ORGANIZER;CN="abc,Some "long" presenter name goes here":mailto:abc,Some%20%22long%22%20presenter%20name%20goes%20here@matterhorn.opencast
This both has encoding issues (quotes within quotes) and also a non-compliant overlong line.

So simplest is just to drop ORGANIZER, as it's not meaningful or useful and can cause problems.

CAs that need the presenter info should get this from the attached episode.xml.