MH-13377 Fix scheduler rrule TimeZone issue
When scheduling mutliple events, the admin UI sends an RRULE with UTC
hours and minutes plus start and end date  to the server.
The server than retrieves the timezone from the capture agent and
adjusts start and end date accordingly, which is fine. However, what was
missing was converting the hour and minute in the RRULE.