MH-13381, Use Organization Identifier in Workflows
The maximum size of the serialized workflow instances stored in the
database is huge, causing an unnecessary performance overhead for the
frequent serialization and deserialization to/from XML. One cause for
this is the large organization data stored multiple times in the XML.

This patch drops the use of the serialized organization which is not
frequently being used anyway and replaces it by referencing the
organization via its unique identifier.

The old organization is still kept as an optional field, ensuring
backward compatibility to old serialized workflows. This means that no
migration is required.

There are more references to organizations in the XML caused by the user
data stored in the XML, but they will be dealt with separately.