MH-12643, Allow worspace to read from asset manager
The workspace will get all files from the asset manager via HTTP, even
if the files are accessible as local files. While this will ensure the
data's integrity as no one is able to accidentally write to files in the
asset manager, this protection is unnecessary for a read-only access.

Additional to the slow HTTP download, the current mechanism also
generates a lot more network trafic since all archived files are first
transferred from the storage server to the admin node (which basically
acts as proxy) and then transferred to the workspace (potentially on
another server).

This means that, for example, if a worker requests a media file, it is
transferred twice over the network interface of the admin node,
transforming it into a bottleneck for performance.

This patch now tests if the asset manager storage is accessible via
filesystem and if it is, will try to read files from there directly. If
it fails to access the file, it will still fall back to the old method
downloading the files.

This means that if e.g. the asset manager is not mounted on a worker,
everything will still work the same way it does right now.

Note that this patch makes the workspace access assets directly via the
file system and it is not necessary to have the asset manager running on
the same node (only the asset manager utils which are on all nodes
anyway). This means, that this will work on worker or presentation nodes
without modifications to what is running there.

Also note that the protection against modification of arcgived data
still holds since no general access is granted to the asset manager's
storage system.

---

This patch is an updated version of pull request #38 which was declined
back then for the reason of the workspace getting asset manager specific
code. This is what has changed since then:

- Those who were complaining back then (and many more) are now *actively
  using the old patch in production* ;-)
- Most asset manager specific code (e.g. resolving the asset manager
  URLs) has been moved into the asset manager utils which are readily
  available on all node types.
- Additional to directly reading files for `workspace::read` requests,
  files for `workspace::get` are copied löcally if possible to avoid
  tunneling all traffic through the admin node.