Add new Workflow Operation: Subtitle Timeshift
With this Workflow Operation we can define the flavor of a subtitle track and the flavor of a video track. The workflow operation will then shift all timestamps of the subtitle file by the duration of the video.

Use Case: Currently if you put a bumper/intro video in front of an already subtitled presenter track, the captions will mismatch. You can use this workflow to shift the times of the subtitle file by the duration of the bumper/intro video and the captions will match again.