Request user data only if necessary
The index service tries to look up users by what is listed in the
presenter field every time an event is created or updated via the
service. But the user information is used only when dealing with
scheduled events. Otherwise, the information is just discarded.

That doesn't really make sense and so, this patch changes the mechanism
for updating events to only get the user information if we are dealing
with a scheduled event.

This is the “safe” option not changing the behavior, although I'm
reasonably sure that the whole mechanism is superfluous and never
actually used for anything. This is supported by checking two production
systems which both gathered no user data all all through this mechanism.
Still, this is already an improvement and far less intrusive.