Treat `filter` parameter elements in admin UI APIs as URL encoded
This allows us to use literal `,` and `:` in filters, cf. #1673,
specifically https://github.com/opencast/opencast/issues/1673#issuecomment-2093704704.

Note that this changes behavior!
Arguably a comma leading to filters being misinterpreted is a bug,
but this "fixes" it by also changing the meaning of certain other filters.
(Think literal `%2C` for example.)

Note also that I had to duplicate the old version of the method used
into the `GroupsEndpoint` of the external API as to not change its behavior.
This makes most (hopefully all) admin UI endpoints' behavior regarding
the `filter` parameter (more) inconsistent with the external API.

Fixing that would be unproportionally more effort in my opinion,
and if I were to do that, I would probably completely revamp
how filters are passed/processed to begin with, because frankly
naicely spliitting a string at `,` was never a good idea.

This way at leaast clicking on a name with a comma in the admin UI
does the right thing. 🤷‍♀️