Fix Amberscript Transcription Jobs Hanging
This patch fixes the problem that AmberScript transcription jobs will
never finish if the event has been deleted between starting and
finishing the transcription.

The code checked if a snapshot exists, skipping the operation and
repeating this on the next run again. An existing event will always have
at least one snapshot, since one is created at ingest or even earlier if
the event was scheduled. If you delete the event, however, there will be
no snapshot, but there will obviously also never be any snapshot. That
is why the jobs will just wait forever.